chore: Adding more consistency in shell scripts
As we are working in a multi os domain (Linux/mac/windows)
we need to take care of different policies on the way shells
are called in shebangs.

`#!/bin bash` do not work on macOs and on certain linux having done
strangely the usrmerge migration.

So I propose to use the more consistent
`#!/usr/bin/env bash`

In the same spirit, I changed one `.bash` extension to `.sh`